Moving ChebyshevAddition to SIE
Temporarily moving this file because I’d like to start working on the
different cases from different spaces. We can move it back later, but
it was failing to load the package with a new function signature:
function
ProductFun{S<:Chebyshev,T,U<:Union(Chebyshev,JacobiWeight{Chebyshev}),V<
:Union(Chebyshev,JacobiWeight{Chebyshev})).

whereas with all of ApproxFun loaded, this can be compiled no problem.