LowRankReform
Reorganize the methods

add safety for LowRankFun((x,y)->0.0);
based on the epsilon of 0 in that precision divided by the type’s
epsilon.

add methods for constants: LowRankFun(10+im);

add Cholesky factorization of a Hermitian positive (negative) definite
cmatrix

Since that is only a property of the function & domains, and not the
spaces, if the spaces don’t match, we first create it with the dx
space, then convert coefficients of the B Funs. This has the advantage
of still only creating half the Funs in the adaptive stage, then
converting in O(n) arithmetic afterwards.

f(x,y) = besselj0(100(y-x))
@time F = LowRankFun(f);
@time G = LowRankFun(f;method=:Cholesky);