fixes Multiplication{F,ConstantSpace,T}
I thinnk this works. The default was adding entries of the matrix of
Multiplication values in the space of f, but from ConstantSpace we only
need the first column (Σ[F] is an operator of the same rank as F).

Not sure how the default Multiplication was interacting with the
Conversion operator from a ConstantSpace. This might be missing
something from either type promotion or array-valued spaces. The tests
passed though.