Change to Space{Domain,Range}
This is a major change, where Space now has parameters
Space{Domain,Range}. That is, a A space is a map from Domain (which is
now more clearly a subset of a Type) to the type Range. For example,
Chebyshev <: Space{Interval{Float64},Float64} as it maps values in
Interval{Float64} to Float64.

Simiularly, VectorSpace{Chebyshev} is now
Space{Interval{Float64},Vector{Float64}} as it maps values in the
interval to float vectors.

Note the coefficients don’t play a role here (yet), so only the
behaviour of the basis matters.