call deploydocs with arguments that Documenter currently supports (#668)
The julia and osname arguments no longer exist.  Those things are
supposed to be specified in travis.yml.

The development branch of ApproxFun is 9 months behind master, so
manual for the development version of ApproxFun should come from
the master branch.  Deploydocs does that by default.

The other arguments to deploydocs are the default values.  Specifying
them explicitly has no effect, except that there is more code for
maintainers to understand.